Add semistandard, some renames
