+++ .gitignore
@@ -1,4 +1,5 @@
+.nyc_output
+node_modules
 
-/node_modules/
+npm-debug.log
 reference
-
