+++ .travis.yml
@@ -1,4 +1,18 @@
+sudo: false
+os:
+  - linux
 language: node_js
 node_js:
-  - "0.11"
   - "0.10"
+  - "0.11"
+  - "0.12"
+  - "4"
+  - "5"
+env:
+  matrix:
+    - TEST_SUITE=test
+matrix:
+  include:
+    - node_js: "4"
+      env: TEST_SUITE=lint
+script: npm run $TEST_SUITE
