+++ test/index.js
@@ -1,10 +1,9 @@
-
 var test = require('tap').test;
-var xorshift = require('./xorshift.js');
+var xorshift = require('../xorshift128+.js');
 
-var reference = require('./reference.json');
+var reference = require('../reference/xorshift128+.json');
 
-function hexview(arr) {
+function hexview (arr) {
   var a = arr[0].toString(16);
   var b = arr[1].toString(16);
 
@@ -14,7 +13,7 @@ function hexview(arr) {
   return (a + b).toUpperCase();
 }
 
-function floatview(d) {
+function floatview (d) {
   // Makes sure that the exponent has two digets like in C-printf
   var s = d.toExponential(20);
   var m = s.match(/^([0-9.]+)e(\+|-)([0-9]+)$/);
@@ -74,7 +73,7 @@ test('default instance', function (t) {
   t.test('random int', function (t) {
     // demand that the 100 first outputs are different
     var obj = Object.create(null);
-    for (var i = 0; i< 100; i++) {
+    for (var i = 0; i < 100; i++) {
       obj[hexview(xorshift.randomint())] = 1;
     }
     t.equal(Object.keys(obj).length, 100);
@@ -84,7 +83,7 @@ test('default instance', function (t) {
   t.test('random double', function (t) {
     // demand that the 100 first outputs are different
     var obj = Object.create(null);
-    for (var i = 0; i< 100; i++) {
+    for (var i = 0; i < 100; i++) {
       var rand = xorshift.random();
       obj[rand.toExponential(20)] = 1;
       t.ok(rand >= 0 && rand < 1, 'random double is [0, 1)');
@@ -99,7 +98,7 @@ test('bad initialization', function (t) {
   t.test('wrong input type', function (t) {
     var error = null;
     try {
-      xorshift.constructor("0102");
+      xorshift.constructor('0102');
     } catch (e) { error = e; }
 
     t.equal(error.name, 'TypeError');
